+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 # 0.3.2
 
  * Fixed: `Copy-GitRepository` intermittently fails when using SSH.
+ * Fixed: `Copy-GitRepository` takes an order of magnitude longer than normal Git because it updates the clone's progress too frequently. It now only updates progress every 1/10th of a second, which has minimal impact on clone speed.
  
 
 # 0.3.1 (30 December 2016)